Add refresh_immv() function The new function refresh_immv(immv_name, with_data) has two argument.  immv_name is incrementally maintainable materialized view's name, and  with_data is an option that is corresponding to WITH [NO] DATA of REFRESH MATERIALIZED VIEW command.  One of differences between IMMVs and normal materialized view refreshed with no data is that IMMVs can be referenced by SELECT although it doesn't return any rows while normal materialized views are not scanable.(#14)
refresh_immv(immv_name, with_data) is a function to refresh IMMV like
 REFRESH MATERIALIZED VIEW command. It has two argument. 
immv_name is incrementally maintainable materialized view's name, and  
with_data is an option that is corresponding to the WITH [NO] DATA option.
When with_data is set false, the IMMV gets unpopulated.

One of differences between IMMVs unpopulated by this function and
normal materialized views unpopulated by REFRESH ... WITH NO DATA
is that such IMMVs can be referenced by SELECT but return no rows, 
while unpopulated materialized views are not scanable. 

The behaviour may be changed in future to raise an error when unpopulated
an IMMV is scanned.